requirements.txt
infi.systray
pyvbox
pywin32